Download manager

Approved-by: Craig van Heerden <craig.vanheerden@entelect.co.za>
Approved-by: Martin Meiring <martin.meiring@entelect.co.za>

+++ ios/media-downloader/RCTMediaDownloaderManager.swift
@@ -30,54 +30,58 @@ class MediaDownloader: RCTEventEmitter {
     
     func prepareDownloader() {
         if (downloadSession == nil) {
-            let backgroundConfiguration = URLSessionConfiguration.background(withIdentifier: "react-native-media-kit-downloader")
+            let backgroundConfiguration = URLSessionConfiguration.background(withIdentifier: "react-native-media-suite-downloader")
             downloadSession = AVAssetDownloadURLSession(configuration: backgroundConfiguration,
                                                         assetDownloadDelegate: self,
                                                         delegateQueue: OperationQueue.main)
         }
     }
     
-    @objc func downloadStream(_ url: String, downloadID: String) {
-        downloadStreamHelper(url: url, downloadID: downloadID)
+    @objc func downloadStream(_ url: String, downloadID: String, title: String, assetArtworkURL: String) {
+        downloadStreamHelper(url: url, downloadID: downloadID, title: title, assetArtworkURL: assetArtworkURL)
     }
     
-    @objc func downloadStreamWithBitrate(_ url: String, downloadID: String, bitRate: NSNumber) {
-        downloadStreamHelper(url: url, downloadID: downloadID, bitRate: bitRate)
+    @objc func downloadStreamWithBitrate(_ url: String, downloadID: String, title: String, assetArtworkURL: String, bitRate: NSNumber) {
+        downloadStreamHelper(url: url, downloadID: downloadID, title: title, assetArtworkURL: assetArtworkURL, bitRate: bitRate)
     }
     
-    func downloadStreamHelper(url: String, downloadID: String, bitRate: NSNumber?=nil) {
+    func downloadStreamHelper(url: String, downloadID: String, title: String, assetArtworkURL: String, bitRate: NSNumber?=nil) {
+        if #available(iOS 10.0, *) {
         
-        self.prepareDownloader()
-        
-        if !(isDownloaded(downloadID: downloadID)) && !(isDownloading(downloadID: downloadID)) {
+            self.prepareDownloader()
+    
+            if !(isDownloaded(downloadID: downloadID)) && !(isDownloading(downloadID: downloadID)) {
             
-            let asset = AVURLAsset(url: URL(string:url)!)
+                let asset = AVURLAsset(url: URL(string:url)!)
             
-            if #available(iOS 10.0, *) {
+                var data: Data?=nil;
+                if (assetArtworkURL != "none") {
+                    let url = URL(string: assetArtworkURL)
+                    data = try? Data(contentsOf: url!)
+                }
                 
                 let downloadTask: AVAssetDownloadTask?
                 if let unwrappedBitRate = bitRate {
-                    downloadTask = downloadSession.makeAssetDownloadTask(asset: asset,
-                                                                         assetTitle: downloadID as String,
-                                                                         assetArtworkData: nil,
-                                                                         options: [AVAssetDownloadTaskMinimumRequiredMediaBitrateKey: unwrappedBitRate])
-                    
+                        downloadTask = downloadSession.makeAssetDownloadTask(asset: asset,
+                                                                             assetTitle: title as String,
+                                                                             assetArtworkData: data,
+                                                                             options: [AVAssetDownloadTaskMinimumRequiredMediaBitrateKey: unwrappedBitRate])
+                        
                 } else {
                     downloadTask = downloadSession.makeAssetDownloadTask(asset: asset,
-                                                                         assetTitle: downloadID as String,
-                                                                         assetArtworkData: nil)
+                                                                             assetTitle: title as String,
+                                                                             assetArtworkData: data)
                 }
                 
                 downloadTask?.taskDescription = downloadID
                 downloadTask?.resume()
                 activeDownloadsMap[downloadID] = downloadTask
-                
             } else {
-                self.sendEvent(withName: "onDownloadError", body:["error" : "Download not supported for iOS versions prior to iOS 10", "errorType" : "NOT_SUPPORTED", "downloadID" : downloadID])
+                self.sendEvent(withName: "onDownloadError", body: ["error" : "The asset is already downloaded", "errorType" : "ALREADY_DOWNLOADED", "downloadID" : downloadID])
                 return
             }
         } else {
-            self.sendEvent(withName: "onDownloadError", body: ["error" : "The asset is already downloaded", "errorType" : "ALREADY_DOWNLOADED", "downloadID" : downloadID])
+            self.sendEvent(withName: "onDownloadError", body:["error" : "Download not supported for iOS versions prior to iOS 10", "errorType" : "NOT_SUPPORTED", "downloadID" : downloadID])
             return
         }
     }
@@ -89,7 +93,7 @@ class MediaDownloader: RCTEventEmitter {
     func isDownloaded(downloadID: String) -> Bool {
         if UserDefaults.standard.object(forKey: downloadID) != nil {
             let baseURL = URL(fileURLWithPath: NSHomeDirectory())
-            let assetURL = baseURL.appendingPathComponent((UserDefaults.standard.url(forKey: downloadID)?.relativeString)!)
+            let assetURL = baseURL.appendingPathComponent((UserDefaults.standard.url(forKey: downloadID)?.path)!)
             if FileManager.default.fileExists(atPath: assetURL.path) {
                 return true
             }
@@ -106,32 +110,68 @@ class MediaDownloader: RCTEventEmitter {
     @objc func deleteDownloadedStream(_ downloadID: String) {
         let userDefaults = UserDefaults.standard
         
+        if isDownloading(downloadID: downloadID) {
+            self.cancelDownload(downloadID)
+            return
+        }
+        
         if isDownloaded(downloadID: downloadID) {
             do {
                 let baseURL = URL(fileURLWithPath: NSHomeDirectory())
-                let assetURL = baseURL.appendingPathComponent((UserDefaults.standard.url(forKey: downloadID)?.relativeString)!)
+                let assetURL = baseURL.appendingPathComponent((UserDefaults.standard.url(forKey: downloadID)?.path)!)
                 try FileManager.default.removeItem(atPath: assetURL.path)
                 
                 userDefaults.removeObject(forKey: downloadID)
+                return
             } catch {
-                self.sendEvent(withName: "onDownloadError", body: ["error" : "An error occured deleting the file: \(error)", "errorType" : "DELETE", "downloadID" : downloadID])
+                self.sendEvent(withName: "onDownloadError", body: ["error" : "An error occured deleting the file: \(error)", "errorType" : "DELETE_FAILED", "downloadID" : downloadID])
+                return
             }
         }
+        self.sendEvent(withName: "onDownloadError", body: ["error" : "Download not found.", "errorType" : "NOT_FOUND", "downloadID" : downloadID])
     }
     
     @objc func cancelDownload(_ downloadID: String) {
+        if (!isDownloading(downloadID: downloadID)) {
+            if (isDownloaded(downloadID: downloadID)) {
+                self.sendEvent(withName: "onDownloadError", body: ["error" : "Finished download cannot be canceled.", "errorType" : "CANCEL_FAILED", "downloadID" : downloadID])
+                return
+            }
+            self.sendEvent(withName: "onDownloadError", body: ["error" : "Download not found.", "errorType" : "NOT_FOUND", "downloadID" : downloadID])
+            return
+        }
         activeDownloadsMap[downloadID]?.cancel()
         print("(\(downloadID)) cancelled")
     }
     
     @objc func pauseDownload(_ downloadID: String) {
+        if (!isDownloading(downloadID: downloadID)) {
+            self.sendEvent(withName: "onDownloadError", body: ["error" : "Download not found.", "errorType" : "NOT_FOUND", "downloadID" : downloadID])
+            return
+        }
+        if (isDownloaded(downloadID: downloadID)) {
+            self.sendEvent(withName: "onDownloadError", body: ["error" : "Finished download cannot be paused.", "errorType" : "PAUSE_FAILED", "downloadID" : downloadID])
+            return
+        }
         activeDownloadsMap[downloadID]?.suspend()
         print("(\(downloadID)) paused")
     }
     
     @objc func resumeDownload(_ downloadID: String) {
+        if (!isDownloading(downloadID: downloadID)) {
+            self.sendEvent(withName: "onDownloadError", body: ["error" : "Download not found.", "errorType" : "NOT_FOUND", "downloadID" : downloadID])
+            return
+        }
+        if (isDownloaded(downloadID: downloadID)) {
+            self.sendEvent(withName: "onDownloadError", body: ["error" : "Finished download cannot be resumed.", "errorType" : "RESUME_FAILED", "downloadID" : downloadID])
+            return
+        }
         activeDownloadsMap[downloadID]?.resume()
         print("(\(downloadID)) resumed")
+    }
+    
+    @objc func setMaxSimultaneousDownloads(_ maxSimultaneousDownloads: NSInteger) {
+        downloadSession.delegateQueue.maxConcurrentOperationCount = maxSimultaneousDownloads;
     }
     
 }
